Move UdevTest class into _utils
